put the Classes into an extra file for more clearness
