+++ app.py
@@ -1,14 +1,13 @@
 import json
 import numpy as np # linear algebr
 import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
-
-from flask import Flask, request
-from functools import reduce
 import os
 import pyrebase
-from random import randint
 import socket
-from textblob import TextBlob
+from flask import Flask, request
+from random import randint
+from UserClass import User, Request
+
 
 app = Flask(__name__)
 
@@ -34,66 +33,6 @@ def start():
   # get_usertypes
   # get_userneeds
   # save_with_analyzed=True
-
-class User(object):
-    """ classifies the attributes of the user and gets the overall sentiment of the users requests """
-    def __init__(self, session, timestamp):
-
-        self.log_list = [Request(session['chat'][log]) for log in session['chat']]
-        print(self.log_list)
-        self.sessionid = timestamp
-        self.session = session
-
-    def get_sentiment_overall(self):
-        request = [req.get_sentiment() for req in self.log_list]
-        sen_array = []
-        for sen in request:
-            sen_array.append(sen[1])
-
-        if len(sen_array) > 1:
-            mean_sen = reduce((lambda x,y: x + y), sen_array) / len(sen_array)
-            return mean_sen
-        else:
-            return sen_array
-        pass
-
-
-class Request(object):
-
-    def __init__(self, log):
-        #print(log['request'])
-        self.id = id
-        self.query = log['request']['query']
-        self.client = log['request']['query']['client']
-        self.location = log['request']['query']['location']
-        self.notmatched = log['request']['query']['notmatched']
-        self.productid = log['request']['query']['productid']
-        self.product_key_words = log['request']['query']['product_key_words']
-        self.question_key_words = log['request']['query']['question_key_words']
-        self.question_words = log['request']['query']['question_words']
-        self.verb = log['request']['query']['verb']
-        self.sentiment = log['request']['sentiment']
-        self.sentiment_type = log['request']['type']
-        self.text = log['request']['text']
-        self.response = log['response']['text']
-
-    def get_sentiment(self):
-        string = self.text
-        polarity = TextBlob(string)
-        pol = polarity.sentiment
-
-        if pol[0] > 0:
-            self.sentiment = pol[0]
-            self.sentiment_type = 'pos'
-        elif pol[0] == 0:
-            self.sentiment = pol[0]
-            self.sentiment_type = 'neu'
-        else:
-            self.sentiment = pol[0]
-            self.sentiment_type = 'neg'
-
-        return (self.sentiment_type, self.sentiment)
-
 
 def get_chatlog_stream(message):
     if message['path'] == '/':
